Refactor - Add NestedQueryBuilder module
